requirements.txt
pyTelegramBotAPI
python-dotenv
requests
pandas
openpyxl
